+++ packages/firebase_core/firebase_core/lib/src/firebase.dart
@@ -25,9 +25,7 @@ class Firebase {
 
   /// Returns a list of all [FirebaseApp] instances that have been created.
   static List<FirebaseApp> get apps {
-    return _delegate.apps
-        .map((app) => FirebaseApp._(app))
-        .toList(growable: false);
+    return _delegate.apps.map(FirebaseApp._).toList(growable: false);
   }
 
   /// Initializes a new [FirebaseApp] instance by [name] and [options] and returns