+++ packages/firebase_core/firebase_core/lib/src/internals.dart
@@ -77,7 +77,7 @@ Object _mapException(
 
 /// Will return a [FirebaseException] from a thrown web error.
 /// Any other errors will be propagated as normal.
-R guard<R>(
+R guardWebExceptions<R>(
   R Function() cb, {
   required String plugin,
   required String Function(String) codeParser,
@@ -88,38 +88,47 @@ R guard<R>(
 
     if (value is Future) {
       return value.catchError(
-        (err) => throw _mapException(
-          err,
-          plugin: plugin,
-          codeParser: codeParser,
-          messageParser: messageParser,
+        (err, stack) => Error.throwWithStackTrace(
+          _mapException(
+            err,
+            plugin: plugin,
+            codeParser: codeParser,
+            messageParser: messageParser,
+          ),
+          stack,
         ),
         test: _testException,
       ) as R;
     } else if (value is Stream) {
       return value.handleError(
-        (err) => throw _mapException(
-          err,
-          plugin: plugin,
-          codeParser: codeParser,
-          messageParser: messageParser,
+        (err, stack) => Error.throwWithStackTrace(
+          _mapException(
+            err,
+            plugin: plugin,
+            codeParser: codeParser,
+            messageParser: messageParser,
+          ),
+          stack,
         ),
         test: _testException,
       ) as R;
     }
 
     return value;
-  } catch (error) {
+  } catch (error, stack) {
     if (!_testException(error)) {
       // Make sure to preserve the stacktrace
       rethrow;
     }
 
-    throw _mapException(
-      error,
-      plugin: plugin,
-      codeParser: codeParser,
-      messageParser: messageParser,
+    Error.throwWithStackTrace(
+      _mapException(
+        error,
+        plugin: plugin,
+        codeParser: codeParser,
+        messageParser: messageParser,
+      ),
+      stack,
     );
   }
 }