+++ packages/firebase_database/firebase_database_platform_interface/lib/src/method_channel/method_channel_on_disconnect.dart
@@ -27,7 +27,7 @@ class MethodChannelOnDisconnect extends OnDisconnectPlatform {
         }),
       );
     } catch (e, s) {
-      throw convertPlatformException(e, s);
+      convertPlatformException(e, s);
     }
   }
 
@@ -45,7 +45,7 @@ class MethodChannelOnDisconnect extends OnDisconnectPlatform {
         ),
       );
     } catch (e, s) {
-      throw convertPlatformException(e, s);
+      convertPlatformException(e, s);
     }
   }
 
@@ -60,7 +60,7 @@ class MethodChannelOnDisconnect extends OnDisconnectPlatform {
         database.getChannelArguments({'path': ref.path}),
       );
     } catch (e, s) {
-      throw convertPlatformException(e, s);
+      convertPlatformException(e, s);
     }
   }
 
@@ -75,7 +75,7 @@ class MethodChannelOnDisconnect extends OnDisconnectPlatform {
         }),
       );
     } catch (e, s) {
-      throw convertPlatformException(e, s);
+      convertPlatformException(e, s);
     }
   }
 }