+++ packages/cloud_firestore_odm/cloud_firestore_odm_generator/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.0-dev.10
+
+ - **FIX**: update all Dart SDK version constraints to Dart >= 2.16.0 (#8184). ([df4a5bab](https://github.com/FirebaseExtended/flutterfire/commit/df4a5bab3c029399b4f257a5dd658d302efe3908))
+
+## 1.0.0-dev.9
+
+ - **FIX**: Use descending in orderBy* (#8159). ([0b7b8811](https://github.com/FirebaseExtended/flutterfire/commit/0b7b88117ac65a0ab164ffcaa0ca7fa69633fcb2))
+
 ## 1.0.0-dev.8
 
  - Update a dependency to the latest release.