+++ packages/firebase_auth/firebase_auth/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.3.9
+
+ - **FIX**: update all Dart SDK version constraints to Dart >= 2.16.0 (#8184). ([df4a5bab](https://github.com/FirebaseExtended/flutterfire/commit/df4a5bab3c029399b4f257a5dd658d302efe3908))
+
+## 3.3.8
+
+ - Update a dependency to the latest release.
+
 ## 3.3.7
 
  - **DOCS**: Update documentation for `currentUser` property to make expectations clearer. (#7843). ([59bb47c2](https://github.com/FirebaseExtended/flutterfire/commit/59bb47c2490fbd641a1fcc26f2f888e8f4f02671))