+++ packages/firebase_auth/firebase_auth_web/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.3.9
+
+ - **FIX**: update all Dart SDK version constraints to Dart >= 2.16.0 (#8184). ([df4a5bab](https://github.com/FirebaseExtended/flutterfire/commit/df4a5bab3c029399b4f257a5dd658d302efe3908))
+
+## 3.3.8
+
+ - Update a dependency to the latest release.
+
 ## 3.3.7
 
  - **FIX**: Add support for`dynamicLinkDomain` property to `ActionCodeSetting` for web. (#7683). ([3b0bf76e](https://github.com/FirebaseExtended/flutterfire/commit/3b0bf76e015c95840b2d38eec7f12c001d3bd47c))