+++ packages/firebase_dynamic_links/firebase_dynamic_links/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 4.0.8
+
+ - **FIX**: update all Dart SDK version constraints to Dart >= 2.16.0 (#8184). ([df4a5bab](https://github.com/FirebaseExtended/flutterfire/commit/df4a5bab3c029399b4f257a5dd658d302efe3908))
+
+## 4.0.7
+
+ - Update a dependency to the latest release.
+
 ## 4.0.6
 
  - **FIX**: Ensure Dynamic link is retrieved from the Intent just once for `getInitialLink()` on Android as per the documentation. (#7743). ([67cc6647](https://github.com/FirebaseExtended/flutterfire/commit/67cc66471046822463f326c05e732313dbaa9560))