+++ packages/firebase_ml_model_downloader/firebase_ml_model_downloader/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.0+8
+
+ - **FIX**: update all Dart SDK version constraints to Dart >= 2.16.0 (#8184). ([df4a5bab](https://github.com/FirebaseExtended/flutterfire/commit/df4a5bab3c029399b4f257a5dd658d302efe3908))
+
+## 0.1.0+7
+
+ - Update a dependency to the latest release.
+
 ## 0.1.0+6
 
  - **FIX**: fixed an issue where macOS builds failed due to bug with missing pod subspec in Firebase SDK (added a workaround until issue fixed upstream). ([acc6afab](https://github.com/FirebaseExtended/flutterfire/commit/acc6afab3e1bc66baa13a35c968df45de36e64bf))