+++ packages/firebase_remote_config/firebase_remote_config_web/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.7
+
+ - **FIX**: update all Dart SDK version constraints to Dart >= 2.16.0 (#8184). ([df4a5bab](https://github.com/FirebaseExtended/flutterfire/commit/df4a5bab3c029399b4f257a5dd658d302efe3908))
+
+## 1.0.6
+
+ - Update a dependency to the latest release.
+
 ## 1.0.5
 
  - Update a dependency to the latest release.